Liohufg

+++ pages/index.tsx
@@ -14,7 +14,7 @@ const Home: NextPage = () => {
 
       <main className={styles.main}>
         <h1 className={styles.title}>
-          Welcome to <a href="https://nextjs.org">Next.js! Long</a>
+          Welcome to <a href="https://nextjs.org">Next.js! Long long</a>
         </h1>
 
         <p className={styles.description}>
